+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `pcb release` now generates a canonical `netlist.json` in the release staging directory
+
 ## [0.3.23] - 2026-01-13
 
 ### Added